requirements.txt
PySide6-Essentials==6.5.1.1
NetworkX==3.1
